Split update_work into three updater files
